32016 Fix typo in T16In32 type

+++ NS32016/32016.h
@@ -321,7 +321,7 @@ typedef union
       uint16_t Upper;
    };
 
-   uint16_t Whole;
+   uint32_t Whole;
 } T16In32;
